% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{augment_edges}
\alias{augment_edges}
\title{Returns a tidy data frame with one row per edge in the original graph}
\usage{
augment_edges(pathway)
}
\arguments{
\item{pathway}{A \code{pathfinder_results} object from, e.g. \code{\link{greedy_search}}.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with the following columns:
\itemize{
\item \code{edge_id} edge index in original graph
\item \code{bundle_id} Bundle id for this edge if it belonged to one
\item \code{times_edge_crossed} Total times this edge was crossed by the path
}
}
\description{
Returns a tidy data frame with one row per edge in the original graph
}
