% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{augment_path}
\alias{augment_path}
\title{Returns a tidy data frame with one row per edge crossing}
\usage{
augment_path(pathway)
}
\arguments{
\item{pathway}{A \code{pathfinder_results} object from, e.g. \code{\link{greedy_search}}.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with the following columns:
\itemize{
\item \code{index} Row number
\item \code{step_id} Step of the pathfinding algorithm
\item \code{edge_id} ID of edge crossed in this step (can be repeated)
\item \code{bundle_id} Id of bundle crossed in this step (\code{NA} if edge is not a bundle)
\item \code{times_edge_crossed} Cumulative times this edge has been crossed since the start of the path
\item \code{times_bundle_crossed} Cumulative times this bundle has been crossed since the start of the path
}
}
\description{
Returns a tidy data frame with one row per edge crossing
}
