% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_attributes.R
\name{first}
\alias{first}
\title{First value in a vector}
\usage{
first(v)
}
\arguments{
\item{v}{A vector.}
}
\value{
The first value of \code{v}.
}
\description{
This is modeled on dplyr's \code{\link[dplyr:first]{first}}. Used as the default edge
attribute combination function in \code{\link[=simplify_topology]{simplify_topology()}}.
}
