% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_interface_points}
\alias{get_interface_points}
\title{Find the vertices in a graph where bundled and non-bundled edges are tangent}
\usage{
get_interface_points(graph, edge_bundles)
}
\arguments{
\item{graph}{An \code{\link{igraph}} object.}

\item{edge_bundles}{A list of integer vectors. Each list item will be
considered one edge bundle, and each vector the edge indices belonging to
that bundle.}
}
\description{
Find the vertices in a graph where bundled and non-bundled edges are tangent
}
