% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graphComponents}
\alias{graphComponents}
\title{Subset all connected components of a graph from node u}
\usage{
graphComponents(graph, v)
}
\arguments{
\item{graph}{A graph object.}

\item{v}{Node id.}
}
\description{
\code{graphComponents} gives a subgraph in which any vertices can be
reached departing from node v by at least one path.
}
\examples{
graph = randomGraph(n=10, .15)
subsetGraph = graphComponents(graph, 4)

}
