% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAdjacencyList.R
\name{makeAdjcList}
\alias{makeAdjcList}
\title{Create a static graph Adjacency list}
\usage{
makeAdjcList(
  min.lon,
  max.lon,
  min.lat,
  max.lat,
  dx,
  dy = NULL,
  n = 1,
  land = NULL
)
}
\arguments{
\item{min.lon}{A value indicating the min longitude of the graph (numeric).}

\item{max.lon}{A value indicating the max longitude of the graph  (numeric)}

\item{min.lat}{A value indicating the min latitude of the graph  (numeric)}

\item{max.lat}{A value indicating the max latitude of the graph  (numeric)}

\item{dx}{Graph resolution in the x direction}

\item{dy}{Graph resolution in the y direction}

\item{n}{number of adjacent cells (a cuales puede ir, 1 solo a las contiguas, 2 a las contijuas de las contiguas...)}

\item{land}{Layer from which the geometries or attributes are queried. This arguments allows
'SpatialPolygons' or 'SpatialPolygonsDataFrame' class objects.}
}
\value{
This function return node and edge adjacency-list
}
\description{
\code{makeAdjcList} function constructs an adjacency list
}
\examples{
\dontrun{
# Graph id scheme
 #          -------------
 #          | 4 | 5 | 6 |
 #          -------------
 #          | 1 | 2 | 3 |
 #          -------------
 }
}
