% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{makeGraph}
\alias{makeGraph}
\title{Constructs a static weighted graph}
\usage{
makeGraph(from, to, weight = NULL, directed = TRUE, coords = NULL)
}
\arguments{
\item{from}{Integer or character vector containing sources nodes IDs.}

\item{to}{Integer or character vector containing target nodes IDs.}

\item{weight}{A non-negative integer or numeric vector containing the edges weights.}

\item{directed}{logical. If FALSE, then all edges are duplicated by inverting 'from' and 'to' nodes (default TRUE).}

\item{coords}{A data.frame or matrix containing all nodes coordinates. Columns order should be 'id', 'x', 'y'. (Optional)}
}
\description{
Constructs a static weighted graph
}
\details{
The \code{makeGraph} function can be used to construct a weighted graph.
}
\examples{
 from = c(1, 2, 3)
 to = c(2, 3, 1)
 weight = runif(3)
 coords = data.frame("id"=1:3, "x"=10:12, "y"=c(2, 3, 5))
 makeGraph(from=from, to=to, weight = weight, coords = coords)

}
