% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_search.R
\name{matrix_search}
\alias{matrix_search}
\title{Greedy search using precomputed matrices}
\usage{
matrix_search(graph, starting_point = 1)
}
\arguments{
\item{graph}{A graph}

\item{starting_point}{Integer. Starting vertex}
}
\description{
Greedy search using precomputed matrices
}
