% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{pathway}
\alias{pathway}
\alias{unique_pathway_edges}
\alias{pathway_steps}
\title{Get pathway data}
\usage{
unique_pathway_edges(pathway)

pathway_steps(pathway)
}
\arguments{
\item{pathway}{A \code{pathfinder_results} object from, e.g. \code{\link{greedy_search}}.}
}
\description{
Get pathway data
}
\section{Functions}{
\itemize{
\item \code{unique_pathway_edges}: Get unique edges visited by pathway

\item \code{pathway_steps}: Get a list of the nodes visited by a pathway
}}

