% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pgh_bundles}
\alias{pgh_bundles}
\title{Pitsburgh bridge bundles}
\format{A list of edge index vectors from \link{pgh_graph}.}
\source{
\href{https://www.openstreetmap.org}{OpenStreetMap}, available under the Open Database License, © OpenStreetMap contributors.
}
\usage{
pgh_bundles
}
\description{
Each vector groups edge indices belonging to the same bridge.
}
\keyword{datasets}
