% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{quiet_messaging}
\alias{quiet_messaging}
\alias{message_end}
\alias{step_status_message}
\alias{message_increase}
\alias{message_crossed}
\alias{message_removed}
\alias{message_none}
\title{Any function to send a message will do nothing if quiet = TRUE}
\usage{
message_end(quiet, epath)

step_status_message(quiet, starting_point, search_set, is_bundle_crossing,
  bundle_id = NULL)

message_increase(quiet, bundle_edges)

message_crossed(quiet, bundles_crossed)

message_removed(quiet, removed_nodes)

message_none(quiet)
}
\description{
Any function to send a message will do nothing if quiet = TRUE
}
