% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{simplify_topology}
\alias{simplify_topology}
\title{Remove all 2-degree nodes from a graph}
\usage{
simplify_topology(graph, edge_attr_comb = list(.default.combiner =
  first), protected_nodes = NULL, progress = interactive())
}
\arguments{
\item{graph}{An undirected \code{\link[igraph:igraph]{igraph}} object}

\item{edge_attr_comb}{Named list mapping edge attribute merging functions.
See Details.}

\item{protected_nodes}{Integer. Indices of nodes that should NOT be removed}

\item{progress}{Boolean. Display progress bar?}
}
\value{
An undirected \code{\link[igraph:igraph]{igraph}}
}
\description{
Rewires \code{graph} to remove all 2-degree nodes. Nodes in the new graph will
retain their original attributes. New edges will inherit attributes from the
cominbation of their parent edges.
}
\details{
Specify this combination by providing a named list of summary functions.
These summary functions must take a vector and return a scalar value of the
same type. For example:

\code{list(distance = sum, edge_id = max, .default.combiner = min)}

The value of \code{.default.combiner} will be used for any edge attributes that
aren't specified elsewhere in the list.
}
