% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{step_status_message}
\alias{step_status_message}
\title{Status message per step}
\usage{
step_status_message(starting_point, search_set, is_bundle_crossing,
  bundle_id = NULL)
}
\description{
Status message per step
}
