% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{tidy}
\alias{tidy}
\title{Returns a tidy data frame with one row per edge bundle}
\usage{
tidy(pathway)
}
\arguments{
\item{pathway}{A \code{pathfinder_results} object from, e.g. \code{\link{greedy_search}}.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with the following columns
- \code{bundle_id}
- \code{times_crossed}
}
\description{
Returns a tidy data frame with one row per edge bundle
}
