% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorators.R
\name{decorate_graph}
\alias{decorate_graph}
\title{Utility function that applies edge and vertex attributes to a graph that will
be used by pathfinder and then later removed.}
\usage{
decorate_graph(graph, edge_bundles, distances)
}
\description{
Utility function that applies edge and vertex attributes to a graph that will
be used by pathfinder and then later removed.
}
