% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{penalize}
\alias{penalize}
\alias{penalize_square}
\alias{penalize_inf}
\title{Edge weight penalty functions}
\usage{
penalize_square(x)

penalize_inf(x)
}
\arguments{
\item{x}{Numeric. Edge distance}
}
\value{
Numeric the same length as \code{x}
}
\description{
Edge weight penalty functions
}
