% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pgh_distances}
\alias{pgh_distances}
\title{Pittsburgh edge distances}
\format{A numeric vector of edge distances for \link{pgh_graph} in meters.}
\source{
\href{https://www.openstreetmap.org}{OpenStreetMap}, available under the Open Database License, © OpenStreetMap contributors.
}
\usage{
pgh_distances
}
\description{
Pittsburgh edge distances
}
\keyword{datasets}
