% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pgh_graph}
\alias{pgh_graph}
\title{Pitsburgh road network}
\format{An \link[igraph:igraph]{igraph::igraph} object containing a sample of street network data from Pittsburgh.}
\source{
\href{https://www.openstreetmap.org}{OpenStreetMap}, available under the Open Database License, © OpenStreetMap contributors.
}
\usage{
pgh_graph
}
\description{
Pitsburgh road network
}
\keyword{datasets}
